Workaround for #332 - iOS 13 Picture Notifications
Workaround for iOS 13, and the requirement for motion sensors to be part of a camera in order for picture notifications.  Creates a dummy motion sensor and switch as part of the camera, and the switch turns the motion sensor on.  The motion sensor is automatically turned off after 5 seconds.

To enable picture notifications, create an Automation in the home app that turns on the dummy switch when the real motion sensor detects motion.

Also includes pull request #309